Correção do professor do tipo primitivo para o tipo Wrapper

+++ Tarefa_Wrapper/src/Tarefa.java
@@ -1,7 +1,7 @@
 public class Tarefa {
     public static void main(String[] args) {
         int Num = 10;
-        long Num1 = Num;
+        Integer Num1 = Num;
         System.out.println(Num1);
 
     }
